% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBGC.R
\name{plot.BGC}
\alias{plot.BGC}
\title{plot.BGC}
\usage{
\method{plot}{BGC}(modelInput = NULL, variable = NULL, ...)
}
\arguments{
\item{modelInput}{The BGC object variable}

\item{variable}{which variable you want to plot}
}
\description{
This function makes possible to use the build in plotting engine in R on BGC opjects
}
\author{
Roland HOLLÓS
}
