% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBGC.R
\name{readBGC}
\alias{readBGC}
\title{readBGC}
\usage{
calibMuso(settings,parameters=NULL, timee='d', debugging=FALSE, logfilename=NULL,
keepEpc=FALSE, export=FALSE, silent=FALSE, aggressive=FALSE, leapYear=FALSE)
}
\arguments{
\item{settings}{RBBGCMuso uses variables that define the entire simulation environment. Those environment variables include the name of the INI files, the name of the meteorology files, the path to the model executable and its file name, the entire output list, the entire output variable matrix, the dependency rules for the EPC parameters etc. Using the runMuso function RBBGCMuso can automatically create those environment variables by inspecting the files in the working directory (this happens through the setupMuso function). It means that by default model setup is performed automatically in the background and the user has nothing to do. With this settings parameter we can force runMuso to skip automatic environment setup as we provide the environment settings to runMuso. In a typical situation the user can skip this option.}

\item{runType}{you can run the modell in a spinup, normal, or a complete mode. Normal is the default.}

\item{inputParams}{The index of the input values.}
}
\value{
No return, outputs are written to file, if it's soil parameter, pleas provide
}
\description{
This function reads the model's output file in a well-structured way. As the result is passed to R, the results can be easily post-processed in R environment.
}
\author{
Roland HOLLÓS
}
